+++ src/sys/tasklog.h
@@ -77,7 +77,7 @@ extern s32 gSYTasklogTaskID;
 
 extern void func_800048D0(SCTaskGfxCallback arg0);
 extern void func_800048F8(Gfx **dl);
-extern void *syTasklogMalloc(u32 size, u32 alignment);
+extern void* syTasklogMalloc(size_t size, u32 alignment);
 extern void func_80004F78(void);
 extern void syTasklogAppendGfxUcodeLoad(Gfx **dlist, u32 ucodeIdx);
 extern void func_800053CC(void);
